+++ docs/conf.py
@@ -13,6 +13,7 @@ from pathlib import Path
 
 from sphinxcontrib import katex
 
+
 HERE = Path(__file__).parent
 sys.path.insert(0, str(HERE / "extensions"))
 sys.path.insert(0, str(HERE / "sphinx_ext_examples"))
@@ -103,6 +104,7 @@ intersphinx_mapping = {
     "scipy": ("https://docs.scipy.org/doc/scipy", None),
     "pandas": ("https://pandas.pydata.org/docs/", None),
     "scanpy": ("https://scanpy.readthedocs.io/en/stable/", None),
+    "pooch": ("https://www.fatiando.org/pooch/latest/", None),
     "pydantic": ("https://pydantic.dev/docs/validation/", None),
 }
 
@@ -136,5 +138,5 @@ katex_prerender = shutil.which(katex.NODEJS_BINARY) is not None
 nitpick_ignore: list[tuple[str, str]] = [
     # If building the documentation fails because of a missing link that is outside your control,
     # you can add an exception to this list.
-    #     ("py:class", "igraph.Graph"),
+    ("py:class", "scverse_misc._deprecated.CallableWithDeprecatedArg"),
 ]
