+++ stubs/pooch/__init__.pyi
@@ -1,6 +1,34 @@
-from .core import Pooch as Pooch
-from .core import create as create
-from .typing import Processor
+from .typing import Downloader, PathInputType, PathType, Processor
+
+def create(
+    path: PathInputType,
+    base_url: str,
+    version: str | None = None,
+    version_dev: str = "master",
+    env: str | None = None,
+    registry: dict[str, str | None] | None = None,
+    urls: dict[str, str] | None = None,
+    retry_if_failed: int = 0,
+    allow_updates: bool | str = True,
+) -> Pooch: ...
+
+class Pooch:
+    def __init__(
+        self,
+        path: PathType,
+        base_url: str,
+        registry: dict[str, str | None] | None = None,
+        urls: dict[str, str] | None = None,
+        retry_if_failed: int = 0,
+        allow_updates: bool = True,
+    ) -> None: ...
+    def fetch(
+        self,
+        fname: str,
+        processor: Processor | None = None,
+        downloader: Downloader | None = None,
+        progressbar: bool = False,
+    ) -> str: ...
 
 class Unzip:
     def __init__(self, extract_dir: str | None = None) -> None: ...
